Extract Avatar from user profile
